move texture manager declaration and implementation in separate files
